+++ app/controllers/dashboard_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class DashboardController < ApplicationController
+  before_action :assign_debit, only: :index
+
   ##
   # Renders the dashboard page.
   #
@@ -11,6 +13,7 @@ class DashboardController < ApplicationController
   def index
     clear_input_history
     @vehicles_present = !current_user.fleet.empty?
+    @mandates_present = @debit.active_mandates.any?
   end
 
   private
@@ -22,5 +25,10 @@ class DashboardController < ApplicationController
     session[:payment_method] = nil
     session[:submission_method] = nil
     session[:new_payment] = nil
+  end
+
+  # Creates an instance of DirectDebit class and assign it to +@debit+ variable
+  def assign_debit
+    @debit = DirectDebit.new(current_user.account_id)
   end
 end
