+++ app/views/dashboard/_direct_debit.html.haml
@@ -1,5 +1,11 @@
 %article.dashboard-total-group.govuk-grid-column-one-third
-  %p.govuk-body{class: 'govuk-!-font-size-16 govuk-!-margin-top-9'}
-    %strong
-      = link_to('Create a Direct Debit mandate', debits_path, id: 'direct-debit')
-  %p.govuk-body Create a Direct Debit mandate to allow BACS payments
+  - if @mandates_present
+    %p.govuk-body{class: 'govuk-!-font-size-16 govuk-!-margin-top-9'}
+      %strong
+        = link_to('Your direct debits', debits_path, id: 'direct-debit')
+    %p.govuk-body View the Direct Debits you’ve created with each zone
+  - else
+    %p.govuk-body{class: 'govuk-!-font-size-16 govuk-!-margin-top-9'}
+      %strong
+        = link_to('Create a Direct Debit mandate', debits_path, id: 'direct-debit')
+    %p.govuk-body Create a Direct Debit mandate to allow BACS payments
