+++ features/fleets.feature
@@ -33,7 +33,7 @@ Feature: Fleets
     Then I should be on the manage vehicles page
     Then I press the Continue
     Then I should see "You must choose an answer"
-    
+
   Scenario: Removing vehicle from the fleet
     When I have vehicles in my fleet
       And I visit the manage vehicles page
@@ -45,7 +45,7 @@ Feature: Fleets
       And I press the Continue
     Then I should be on the manage vehicles page
       And I should have deleted the vehicle
-    
+
   Scenario: Backend API is unavailable
     When Fleet backend API is unavailable
       And I visit the manage vehicles page
