+++ features/step_definitions/navigate_pages_steps.rb
@@ -2,10 +2,12 @@
 
 When('I navigate to a Dashboard page') do
   mock_vehicles_in_fleet
+  mock_debits('active_mandates')
   visit dashboard_path
 end
 
 When('I navigate to a Dashboard page with empty fleets') do
   mock_empty_fleet
+  mock_debits('inactive_mandates')
   visit dashboard_path
 end
