+++ features/step_definitions/payments_steps.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 When('I visit the make payment page') do
+  mock_debits
   login_user
   visit payments_path
 end
@@ -38,6 +39,7 @@ Then('I should be on the Charge details page') do
 end
 
 And('I want to confirm my payment') do
+  mock_debits
   mock_requests_to_payments_api_with(return_url: result_payments_path)
 end
