+++ features/vehicles.feature
@@ -49,4 +49,3 @@ Feature: Vehicles
       And I press the Confirm
     Then I should be on the details page
       And I should see "You must choose an answer" 2 times
-
