+++ spec/fixtures/responses/debits/inactive_mandates.json
@@ -1,28 +1,28 @@
 {
   "cleanAirZones": [
     {
-       "cazId": "87c17807-d6ec-4128-9174-9ce66cc5ed73",
-       "cazName": "Leeds",
-       "mandates": [
-         {
-           "id": "5cd7441d-766f-48ff-b8ad-1809586fea37",
-           "status": "cancelled"
-         }
-       ]
+      "cazId": "87c17807-d6ec-4128-9174-9ce66cc5ed73",
+      "cazName": "Leeds",
+      "mandates": [
+        {
+          "id": "5cd7441d-766f-48ff-b8ad-1809586fea37",
+          "status": "cancelled"
+        }
+      ]
     },
     {
-       "cazId": "90fa62ca-c9dc-48b3-81c2-66b6f0d28b38",
-       "cazName": "Birmingham",
-       "mandates": [
-         {
-           "id": "09b72307-e56a-4b61-802a-03eee58748ba",
-           "status": "failed"
-         },
-         {
-           "id": "45725067-ae01-4c68-9884-8aac0f5c320b",
-           "status": "cancelled"
-         }
-       ]
+      "cazId": "90fa62ca-c9dc-48b3-81c2-66b6f0d28b38",
+      "cazName": "Birmingham",
+      "mandates": [
+        {
+          "id": "09b72307-e56a-4b61-802a-03eee58748ba",
+          "status": "failed"
+        },
+        {
+          "id": "45725067-ae01-4c68-9884-8aac0f5c320b",
+          "status": "cancelled"
+        }
+      ]
     }
   ]
 }
